Refactor RN Tester url parser
